Moving everything to src directory
