+++ src/instrumentserver/deployment/dashboard.json
