+++ src/instrumentserver/gui/base_instrument.py
