+++ src/instrumentserver/gui/instruments.py
