+++ src/instrumentserver/resource.qrc
