+++ src/instrumentserver/schemas/parameters.json
